+++ sources/Interop/Windows/other/helper-types/AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE.cs
@@ -51,11 +51,11 @@ namespace TerraFX.Interop
 
         public static explicit operator AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE(long value) => new AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE((nint)(value));
 
-        public static implicit operator long(AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE value) => (long)(value.Value);
+        public static implicit operator long(AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE(nint value) => new AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE((nint)(value));
 
-        public static implicit operator nint(AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE value) => (nint)(value.Value);
+        public static implicit operator nint(AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE(sbyte value) => new AUTHZ_CAP_CHANGE_SUBSCRIPTION_HANDLE((nint)(value));
