+++ sources/Interop/Windows/other/helper-types/AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE.cs
@@ -51,11 +51,11 @@ namespace TerraFX.Interop
 
         public static explicit operator AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE(long value) => new AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE((nint)(value));
 
-        public static implicit operator long(AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE value) => (long)(value.Value);
+        public static implicit operator long(AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE(nint value) => new AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE((nint)(value));
 
-        public static implicit operator nint(AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE value) => (nint)(value.Value);
+        public static implicit operator nint(AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE value) => value.Value;
 
         public static explicit operator AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE(sbyte value) => new AUTHZ_SECURITY_EVENT_PROVIDER_HANDLE((nint)(value));
