+++ sources/Interop/Windows/other/helper-types/CIMFS_IMAGE_HANDLE.cs
@@ -51,11 +51,11 @@ namespace TerraFX.Interop
 
         public static explicit operator CIMFS_IMAGE_HANDLE(long value) => new CIMFS_IMAGE_HANDLE((nint)(value));
 
-        public static implicit operator long(CIMFS_IMAGE_HANDLE value) => (long)(value.Value);
+        public static implicit operator long(CIMFS_IMAGE_HANDLE value) => value.Value;
 
         public static explicit operator CIMFS_IMAGE_HANDLE(nint value) => new CIMFS_IMAGE_HANDLE((nint)(value));
 
-        public static implicit operator nint(CIMFS_IMAGE_HANDLE value) => (nint)(value.Value);
+        public static implicit operator nint(CIMFS_IMAGE_HANDLE value) => value.Value;
 
         public static explicit operator CIMFS_IMAGE_HANDLE(sbyte value) => new CIMFS_IMAGE_HANDLE((nint)(value));
